revalidateLinkRecommendations: Make it possible to revalidate based on score
Why:
The link recommendation service provides score for each suggestion,
ranging from 0 to 1. We recently (T316079) changed the minimum
score from 0.5 to 0.6, but not all tasks in the pool have
a score of at least 0.6.

What:
Make it possible to revalidate only link recommendations
with links below the new threshold, via the --scoreLessThan
parameter.

Bug: T316079
Change-Id: Ie1fb241aac209aa11e44e0b716ad4b374c9e4482