WelcomeSurvey: Rename exp2_target_specialpage to control
Why:
The `exp2_target_specialpage` name made sense when Growth
was actively A/B testing the WelcomeSurvey. We use it now,
because it was the best performing name, but in practice,
`exp2_target_specialpage` now represents the new default,
aka the control group for any future experiments (such as
T342353 or T303240).

What:
Rename the exp2_target_specialpage WelcomeSurvey experiment
group to control, to better represent its purpose.

Bug: T344262
Change-Id: If4deaff134d53993c2c31179e7697191f62155b8