LinkRecommendationUpdater: Load link-recommendation even if disabled
Why:
Wiki admins are capable of disablink link-recommendation
from Community configuration. Sometimes, fixing the issues
they report involves running server scripts like
revalidateLinkRecommendations.php, which rely on
LinkRecommendationUpdater. Unfortunately, as of now, it is not
possible to run LinkRecommendationUpdater without link-recommendation
enabled.

It is possible to workaround this limitation, but it is much easier
to just have LinkRecommendationUpdater (and by extension,
revalidateLinkRecommendations.php) runnable without link-recommendation
enabled in Community configuration. Running that script only affects
the task pool, not what the newcomers see.

What:
Check getDisabledTaskTypes() in LinkRecommendationUpdater in addition
to getTaskTypes().

Note:
For now, this can be workarounded by temporarily disabling
wgGELinkRecommendationsFrontendEnabled and enabling the task
type in MediaWiki:NewcomerTasks.json.

Bug: T344343
Change-Id: Id26d5f558c8d659c632dea20d3bb9267d5a564b6