Remove support for wgWelcomeSurveyEnableWithHomepage
Why:
wgWelcomeSurveyEnableWithHomepage disables the experimental
group configuration variable and always uses the default
survey group (control, as of today). This variable was useful
initially, when enwiki wasn't on 100% of Growth features,
to ensure the Welcome survey only shows when Homepage
does (otherwise, it doesn't really have a purpose).

Today, English Wikipedia is on 100% of Growth features and as such,
the configuration property doesn't have any useful purpose
anymore. It is still set to true presumably for legacy reasons,
as having it on didn't have any issues up until now.

Now, Growth wants to deploy an additional survey
group to enwiki (see task) for user testing
opt in checkbox, but wgWelcomeSurveyEnableWithHomepage
makes it impossible for that group to be actually
assigned to the users.

See also I3b2dc52158502144430afe55314620f727132838.

What:
Remove support for wgWelcomeSurveyEnableWithHomepage

Bug: T344619
Change-Id: I4a947e89ed74ba77cddd80bc150a4eab30864462