GrowthTasksApi: make lookahead size configurable
Introduce GEApiQueryGrowthTasksLookaheadSize which will
take precedence over DEFAULT_LOOKAHEAD_SIZE constant if
different from null.

This is to experiment with lookahead increases to
avoid counter drops from the number of tasks available
in ES and the real number of tasks returned to the user
after applying link and protection filters.

Bug: T325990
Change-Id: I84c04769cc6125001fd7f6129bdafa11961a858a