WelcomeSurvey: Update privacy links
Why:
The old link included in the Welcome survey is out of date.
WMF Legal passed recommendations to us for implementation.

The footer (privacy-info) is automatically added to the questions
if omitted by configuration, as it is a Legal-mandated piece
that shouldn't be ommitted. It doesn't make sense to have
to think about it each time we add a survey group.

What:
* Remove sidebar link per WMF Legal and Kirsten
* Stop passing privacy policy link to confirmation messages
  (it is not included in the user-facing message)
* Add a footer to the Welcome survey, to point to the
  survey-specific WMF Privacy statement.

Bug: T343682
Change-Id: I8658a980695641d5586e4e0d7838eee78fc80d1a