Welcome survey: Add generic user-research question
Why:
WMF Legal confirmed the research-specific privacy statement
doesn't need to be included in the Welcome survey, but it can
be included only in the "reaching out" email sent by whoever
runs the research study. As a result, we can include a generic
user-research question in the survey and use it for any user
research studies we need to get newcomers for.

What:
Add generic user-research question to the Welcome survey.

Bug: T342353
Bug: T343682
Change-Id: Ifb21f99d02a128ce54f5005f26a70c9ed0548178