WikiPageConfigWriter: Run EditFilterMergedContentHook
Why:
The AbuseFilter extension depends on the EditFilterMergedContentHook
to validate incoming edits and act on them as appropriate.
WikiPageConfigWriter ignores most permission-related checks, because
we want to allow non-admins to enroll as mentors when appropriate.

However, this does not justify not running the EditFilterMergedContentHook,
as that could be used to implement finer restrictions than what is
readily available.

What:
Make WikiPageCOnfigWriter run the EditFilterMergedContentHook. While
at it, clean up unnecessary @var declarations in the factory class.
This also removes the $allowList param, as it is not read by anything.

Bug: T346235
Change-Id: Ibeb64f242b5e80ae5b8f6493fa27ed57b9557d09