Structured mentor list: Make "no mentees" a proper weight
Why:
Originally, there were two list of mentors (one intended
for users that are assigned to all newcomers automatically,
the second one for users that should be only able to claim
mentees). This was represented in the structured mentor list
via a $autoAssigned boolean.

After that, the weighting system was added, which introduces
a numberic weight. This can be easily used to store
the autoassigned attribute as well, but it didn't happen.

Since the difference between weight and autoassigned status
is not really clear to the mentors (as our end users), this
is all exposed within a single form field. Mixing the regular
weights with a pseudo weight (which is mapped to $autoAssigned false)
causes issues, such as the one reported as T347024.

We still need to respect autoassigned as a possible
in GrowthMentors.json and treat it accordincally. This is
because the wiki list won't be automatically updated by this
patch. The key is removed on every mentor-taken action, so
we can either wait for it to clear by itself, or if we want
to speed things up, we can write a maint script for conversion.
I think waiting for natural cleanup is sufficient.

What:
* Add WEIGHT_NONE to mean "no mentees should be assigned to me"
* If `automaticallyAssigned` is set to false, force the weight
  to WEIGHT_NONE, for backwards compatibility.
* Refactor StructuredMentorProvider to only parse GrowthMentors.json
  on a single place (needed for the weight magic mentioned in the previous
  point to work correctly).
* Formatting

Bug: T347157
Bug: T347024
Change-Id: Ife9ac018a11026dbf9dc3cd42b8d64af5157a12c
(cherry picked from commit fa426f25b858e4b0e209fb0d2be8ce7971ec47c7)