Fix phan for GrowthExperiments
WelcomeSurveyLogger::logInteraction likely did not work
as expected, as getFuzzyBool never returns null; it can
only return a false.

In HomepageHooks::onAuthChangeFormFields, do not condition
around WebRequest::getInt not returning a null, as it can
never return null (when a non-int value is passed, getInt
returns a zero).

In ExportWelcomeSurveyMailingListData, add IDatabase::addQuotes
to suppress the Phan errors. The value appears to be safe, but
adding quotes is both the best practice and cannot do harm.

Bug: T347571
Change-Id: I813e3c26ea68f75c3911971aee7000d56022d896