ExpensiveUserImpact: Do not pass null to DateTime constructor
Why:
When there are no pageview data, then
ExpensiveUserImpact::isPageViewDataStale() fails to find
the latest datetime for pageview data and as such, it sends
null to DateTime::__construct. This means two things:

* as of PHP8.1, this yields a deprecation warning, as
  DateTime::__construct no longer accepts null as timestamp,
* user impact objects with no pageview data are always considered
  not stale (and as such, it is not regenerated by UserImpactHandler
  and friends).

Both behaviors are incorrect. I think we should be attempting to
recalculate user impact objects if there are no pageview data at
all, because that would happen when the user has no edits (and thus
no pages to show in the pageviews analysis). Recalculating user impact
should be cheap in this case, even though it involves pageview data,
and I can't find a compelling reason not to do that.

Even if we decide that not recalculating no-pageview user impacts
is the correct behavior, it should be explicitly recorded in the code,
both to avoid the deprecation warning and to make intentions more clear.

What:
Add explicit branch for no-pageviews scenario to
ExpensiveUserImpact::isPageViewDataStale.

Bug: T349293
Change-Id: I843e1a443b12c926d5f03719a605c7337554ac38