UserImpact: Leave total editcount uncapped
Why:
More tenured users generally have more than 1k of edits. Because of
how long Growth features are a thing, some tenured editors have newcomer
homepage enabled, and are unhappy when the editcount is capped at 999+,
since they cannot keep updates of it.

The cap has two main purposes:
    * some metrics take long time to execute, and a cap can limit
      the maximum time we spend processing data,
    * we do not have unlimited space for displaying numbers, and
      practically speaking, only numbers up to a certain value
      can be reasonably displayed in the space we have available.

The first reason is not important for the user editcount, since
that is available in `user_editcount` field of the `user` table,
and is always readily available. It used to be important previously,
when we only counted mainspace edits in this field (which is an information
that is not readily available).

The second reason is irrelevant as well, since the space designated to
fit the total editcount can easily fit numbers amounting billions,
which is more than what we have to display for any current user,
and for the foreseeable future.

The patch also changes copies for (i) icon for other numbers to include
language about limitting, which is used to inform users what numbers
they're actually seeing.

What:
* Remove the cap for total editcount information
* Rework ComputedUserImpactLookup to use UserEditTracker to get
  the editcount.
* Remove UserOptionsLookup, since that is not used for anything ATM.
* Change (i) messages in en.json (qqq.json).

Bug: T338174
Change-Id: I40ad511694316d54e007f91913af093ecb138de1