changeWikiConfig: Add --touch option
Why:
To reflect config serialization, it is sometimes needed
to make a no-op edit to the config page (that updates it
to the new format). changeWikiConfig.php is an existing
script to make changes to on-wiki config, so adding it
there.

Note --touch only supports GrowthMentors.json as of now.
This is because the way to perform a no-op edit that also
notices possible serialization changes differs for each
config page, and for NewcomerTasks.json / GrowthExperimentsConfig.json,
it would require calling SpecialEditGrowthConfig in that way.
The nice solution would require factoring the serialization
logic out of the special class, which is a non-trivial task,
which makes a little sense to take on now given we're in the middle
of engineering CC2.0.

What:
Add --touch flag to changeWikiConfig

Bug: T347157
Change-Id: I7e41de3b65677618c3817ace0b42e966d0c4ec8d