AddImageSubmissionHandler: Invalidate accepted suggestions
Why:
Prior to this patch, AddImageSubmissionHandler invalidated only
(certain) rejected suggestions. This only makes difference when
the accepting edit is reverted by another user, as then the article
is in unillustrated state with no feedback passed to the pipeline,
resulting in the suggestion staying in the task pool, without anything
removing it.

What:
Fix the condition in AddImageSubmissionHandler::handle, so that
it invalidates accepted submissions as well.

Note:
This should ideally include a regression test. Unfortunately,
AddImageSubmissionHandler::handle is not covered at all, and
adding an unit test would be unreasonably difficult, as it makes
use of the global state via calls like `new ManualLogEntry` and similar.

Bug: T350598
Change-Id: Ieb8988dea6399454469a4c67134417c0d1888dc9