SpecialEditGrowthConfig: Set min to zero for levelling up config
Why:
For both GELevelingUpGetStartedMaxTotalEdits and
GELevelingUpKeepGoingNotificationThresholds, the value of zero
can be used to disable the functionality. While reviewing
Cyndy's patch for levelling up, I missed that in the specifications,
and I mistakenly suggested the minimum for
GELevelingUpKeepGoingNotificationThresholds[1] (the maximum) to be set
to the value of the first index (the minimum), since it makes sense
for min <= max. However, violating this makes sense for disabling
the notifications altogether.

What:
Set the minimum for GELevelingUpKeepGoingNotificationThresholds[1]
to be zero, to support communities disabling the notification.

While at it, set minimum for the other value to zero as well, for
the same reason.

Bug: T343288
Change-Id: If524f2fcbb85a9f11bf95b7e2d6ce3950328f41e