Section images: on mobile use overlay width for setting the image width
Why: something changed within editor initialization on mobile
making the width computation of the surface return 0. Pressumably
the changes from T239676 showing a switch to source editor
on slow connections

What: use the overlay element to get a width value. On phones
the overlay and surface width are the same and on tablet devices
it is capped to 500px.

Bug: T352208
Change-Id: Ib7512f90772ce12744e4222e740ef2a837e148ed