MentorManager: Do not assign a mentor to opted-out users
Why:
When an user opts out of mentorship,
MentorPageMentorManager::setMentorshipStateForUser drops
the mentor/mentee relationship. However, when something
queried for the mentor, MentorManager would happily assign
a new one.

According to local testing, that something can be
the Help panel, which checks for mentorship state on the UI
layer, but when generating wgGEHelpPanelMentorData, it relies
on MentorManager providing accurate data (which is reasonable).

We can avoid this by making MentorManager to check the mentorship
state before checking the mentor.

To ensure the false assignment is not present in dumps,
MentorManager calls MentorStore::dropMenteeRelationship whenever
it detects an assignment which should not be present.

What:
* Add a check to MentorManager::getMentorForUser that the user
  is not opted out, and do not assign a mentor if they are.
* Add a test for this business rule to MentorPageMentorManagerTest

Bug: T351415
Change-Id: I87da89913a17b070930645aebab710ae4bc0941f