Add GEUseCommunityConfigurationExtension feature flag
Why:
The Growth team works on the CommunityConfiguration extension,
which re-implements the ability for administrators to edit
MediaWiki configuration on-wiki in a different way.

This patch allows developers to switch from the GE-specific
implementation of community configuration (so-called CC1.0)
into the new dedicated implementation that CommunityConfiguration
represents.

The patch is meant to make GrowthExperiments stop use any of its
CC1.0 implementation if GEUseCommunityConfigurationExtension is true,
including the editing form.

Notes:
CommunityConfiguration expects client extensions to use a single
Config object (=the one provided by CommunityConfiguration)
for all config needs. This enables client extensions to seamlessly
switch between configurable and non-configurable fields (it is a matter
of changing the scmea), without having to review their usage across their
whole repository. To fully benefit from this expectation, we need to change
how we approach config within GrowthExperiments.

What:
* Introduce `GEUseCommunityConfigurationExtension` config flag, which
  is false by default.
* Create `GrowthExperimentsCommunityConfig`, which either returns
  the `GrowthExperimentsMultiConfig` service or the CommunityConfiguration
  provided service, depending on the feature flag.
* Switch SpecialEditGrowthConfig's registration to a conditional one,
  based on the feature flag value.
* Move two schemas from CommunityConfigurationExample to GrowthExperiments

Bug: T359124
Depends-On: If9db50b8ba8dc835a7d5e1c49de83e7ff2941c01
Change-Id: I2b6e3436c08100442baf4ed65582d0bc17fd449c