Add SuggestedEditsSchema
Immediately after merging the patch,
Special:CommunityConfiguration/GrowthSuggestedEdits will work,
but it will look very broken. This is because of T363477 and
T362098, which will make the form look (mostly) normally.

The form should now look like https://ctrlv.cz/1zNn, and
the config for the three structured tasks should work
as expected.

Bug: T360471
Depends-On: I6b6470ac05316f475660294740fecb4e35296229
Change-Id: I0a25dd94c5f88b2a887dc7879623dd9e3219e2c8