Only treat first colon as a delimeter in onFormatAutocomments
The autocomments have the following format:
message-key:param1|param2|param...|paramN

Prior to this patch, we used explode( ':', $comment )
to parse the comment into the message key and param array.
However, this does not work if there is a colon in one
of the params.

Resolve this issue by using explode( ':', $comment, 2 ),
which only uses the first colon as a delimeter, and ignores
any other colons.

NOTE: The same problem will happen when using a pipe character.
This is less critical, as pipe is less commonly used (and it has
special meaning in wikitext already, so easier to explain). This
is left for a future discussion and/or fix.

Bug: T360846
Change-Id: I5d0974f4e2154ae41cce6629c8e1d19e166a2913