Help panel: Do not break when CommunityConfiguration is used
`GEHelpPanelLinks` is represented as associative arrays in
CC1.0, but in CC2.0, it is returned as a stdClass. Fix
that by adding a casing.

Also remedy PHP Warnings by renaming `label` to `text`,
which is expected in HelpPanel class. Otherwise,
the following notice would be displayed:

Warning: Undefined array key "text" in
/var/www/html/w/extensions/GrowthExperiments/includes/HelpPanel.php
on line 68.

Bug: T363459
Change-Id: I1507dcb6739cb22c5cf049d19b7e3bad8db2c0da