Community configuration: Fix labels for Suggested edits
Prior to this patch,
Special:CommunityConfiguration/GrowthSuggestedEdits
had wrong labels in two places:

* `disabled` checkboxes were unlabeled in all places
   except copyediting
*  Add a link fields were unlabeled

In addition to this, two disabled labels were present,
both for the field and the control, which is inconsistent
with how we decided to represent booleans (cf. Mentorship).

This patch updates the labels for the form to meet expectations.

Bug: T360471
Change-Id: Ib92599555204966709c057a648f840838ebf543a