Config: add maintenance script to migrate to CC2.0
The script contains an allow list of provider ids
to migrate.

When using the --dry-run option the script will
collect the provider/schema informed configuration
options and use the provider validator to check
for the config validity.

On standard run  validity of the collected
configs is also checked and if passed, the new
configuration is stored in the wiki page informed
by the provider.

Note: config options showing as untouched are
most probably legacy options that have not
been removed from the wiki config. Consider
removing them.
Note: config options showing as missing are most
probably options that belong to a feature is not
enabled in the wiki. These need to be reviewed
when running the script.

GEHomepageManualAssignmentMentorsList
GEHomepageMentorsList

Bug: T359038
Depends-On: I1ec7d9f9b67c47955e94ad7f9bec5576bc6f9c8e
Change-Id: Ida92d0e94ae49cc768736fc29e129ca6bc6adf08