Fix package module error message generation
The user session is not available in package modules, and trying to use
that fails with `Sessions are disabled for load entry point`. To fix
that we have to use the \MediaWiki\ResourceLoader\Context::msg method.

See https://www.mediawiki.org/wiki/ResourceLoader/Package_files#Common_pitfall:_using_i18n_messages_in_generated_content_callbacks

Adjust the test because this is now actually treated as a message.

Bug: T365653
Change-Id: Ic1fb83979cc2280a5191fa2842b0e0d9c6659485