Fix PHP Notice: Undefined Index <username>
If a username only consists of digits, then setting it as the key of an
array will set it as a _numeric_ key. MySQL/MariaDB in turn does some
strange type conversion when it receives a numeric value in a where
clause for a field that contains strings. That lead to the query for the
username `555` also returning the usernames `555_É_GAY`, `555éGAYSAÇO`,
and `555éViado`.

Bug: T355344
Change-Id: I9493d67c5cb5d6088723c64950f57ff6512ba528