Fix PHPNotice about index `registration` not being set
This now logs an error with the user id, so that we can follow-up on the
underlying issue. If that reveals that a user with no registration data
is valid, then this logging can be later removed as well.

Based on https://www.mediawiki.org/wiki/Manual:User_table#user_registration
the `registration` field should always have been set except for very old users.
Maybe we just need to run the maintenance script mentioned there for
another wiki.

Bug: T355344
Change-Id: I95494b0cd96e24fa1bab66bf843e5f927b5f9c16