Fix and simplify partially bogus namespaces in tests/phpunit/
Result of running a search-and-replace for `^namespace .*$` in
tests/phpunit/integration and /unit and replacing it with the respective
desired namespace, plus some manual fixes. Using different namespaces
for integration and unit tests allows us to have the same test class in
both directories.

Notes about the manual fixes:

\GrowthExperiments\Tests\Unit\RemoteSearchTaskSuggesterTest::exactlyBetween:
The local package definition no longer defines phpunit in the first
place, so this workaround is no longer needed.

Renamed one class that had needlessly "Integration" in the class/file
name. That's what namespaces are for.

Co-Authored-By: Thiemo Kreuz (WMDE) <thiemo.kreuz@wikimedia.de>
Co-Authored-By: Michael Große <mgrosse@wikimedia.org>
Change-Id: I19d42ff5c00fc471304ea142b91dffa2fc52b56e