Fix broken setMwGlobals calls in tests
setMwGlobals() expects the variable names with the "wg…" prefix,
while overrideConfigValues() enforces it internally. Because the
variable name was not complete this particular setMwGlobals call was
effectively dead code that had no effect on the test.

The FIXME was added in I5b0d664. I don't understand it fully. But it
looks like the broken setMwGlobals was the reason, not what's said
in the comment.

Change-Id: I2d5af543c1aa1f08454926f73a26c6f88617f75b