Hide unavailable features from CommunityConfiguration
The patch hides Personalized praise configuration when
Personalized praise is not available on the wiki. It also hides
configuration for each of the three structured tasks,
when the task is not available on a wiki.

This intentionally adds a new hook class dedicated
for CommunityConfiguration hooks, because GrowthExperiments
needs to run without CommunityConfiguration extension being
available, which is not possible when mixing core and extension
hooks in one handler. See VisualEditorHooks as an example (which
was introduced before VE was a hard dependency for GrowthExperiments).

Bug: T367330
Depends-On: Ia0282d136bff06aa4306027aa7f30fbb4c7eb6f0
Change-Id: Ia30b55ca0ae553730c8139c34d392b543e9cb442