MenteeOverviewDataUpdater: Call autoReconfigure() when waiting for replication
DB-related settings sometime change (when maintenance is due),
and GrowthExperiments should respect the new configuration
whenever possible. Prior to this patch, MenteeOverviewDataUpdater
(called by updateMenteeData.php) was calling LBFactory::waitForReplication,
but did not include the call to LBFactory::autoReconfigure(), which means
old DB config was used for the whole duration of the script.

On bigger wikis, updateMenteeData can take a couple of hours,
so it is important to refresh the config fairly often.
LBFactory::autoReconfigure recommends to call
autoReconfigure() after each batch; add the call
to the same place in which we handle replication lag.

Bug: T367764
Change-Id: I2f8a1f8a23e48c76909cde1e3a6a806da4116f7f