Only log error if both registration and last_edit keys are not set
The registration timestamp can legitimately be null as early mediawiki
did not track the registration date and the backfill might have been
occasionally incomplete.

NOTE: the original error, T355344, was triggered because the registration
key in the array was  **missing** not merely `null`, implying that a row
for a user with that id was not in the users table. This logging is in
principle still more permissive than that, because both last_edit and
registration being merely `null` is also an error condition that we want
to know about.

Bug: T366761
Change-Id: I3087e9df9f46e58323ebbf8a4858f79165b80e4e