Fix inconsistent singular in …-impact-scores-thanks-info-text
Compare e.g. with the message directly below. While I think it's nice
to make it clear that a thanks notification is always only or a single
edit you made, you can receive any number of thanks notifications, by
any user.

It's also much easier to avoid possibly weird gender constructs in
translations like German when we can refer to a group of unknown
contributors instead of a single contributor with unknown gender.

Bug: T368249
Change-Id: I5af2ae04ce20ac056bccb26b001c176860e85b90