Homepage: don't show edit count in mobile no structured tasks fallback
Actually getting the edit count from the API is a very high performance
overhead that recently got much worse and was the main source of latency
for the homepage (see T368405).

This change does the first step by changing the text to longer contain
the edit count as defined in the task, This change focusses on being
small and quickly mergable.

Follow-up changes will include adjusting the image, removing the now
unused code to get the edit count from the Analytics Query Service, and
add tracking.

Bug: T368750
Change-Id: Ic4053819d0387a2af3fdfff2c6b873c4f4761d99