Suggested edits: track how often a user's task queue is error/empty/ok
This serves two purposes:
* be able to prioritize further work on the no-tasks mobile preview
  widget based on data for how many people actually see it
* be able to notice problems from the queue running empty in some wiki
  unusually often

I'm intentionally opting for using the static logger factory over
injecting it as a dependency, because this should never be triggered
aside from a future programming mistake. Alternatively, we could throw
an InvalidArgumentException here, but that feels too disruptive should
such mistake ever occur and make it past code-review. This is only
tracking statistics, so maybe there's no need to throw an uncaught
exception.

Bug: T368750
Change-Id: I41c1f3c287c570987b033dbb8eed49633c32912b