Do not consider Levelling up enabled with SuggestedEdits disabled
Levelling up does not work and make any sense
if Suggested edits are disabled. Check whether
they are enabled before considering Levelling up
enabled. To make it easier, move the checks
into a central place, the LevellingUpManager.

A per-user version of isEnabled is in LevellingUpHooks.
I consider the manager a more appropriate place,
as the manager is responsible for determining
whether the notification should be sent in individual
cases, and determining whether the feature is enabled
is a subset of that decision making.

A follow-up moves the per-user version to the manager,
so that everything is in a single place.

Bug: T369603
Change-Id: I1b6efda19a07248328cc7a230ed354b44cb467b2