Homepage: track total Suggested edit queue views to calculate rates later
While we technically already have the total by ok+empty+error, doing
that calculation in Grafana seems not possible if we want to use the
result in a denominator so that in end we have a panel with the error
rate and empty rate per wiki. Therefore, this now explicitly tracks the
total too.

Bug: T368750
Change-Id: I94c985470cdf0e31a41fc4a3a4c0319e01ff50cf