refreshLinkRecommendations: Work even when link-recommendation is disabled
Why:
refreshLinkRecommendation uses the same TaskSuggester system
as Special:Homepage. For that reason, it currently does not
work when Add Link is disabled via Community configuration.

Because the TaskSuggester system is used in the Homepage
as well, it is not possible to query disabled task types
in it as well. However, it is possible to inject
a custom configuration loader to the TaskSuggester.

This can be used to pretend Add Link is enabled whenever
the refreshLinkRecommendation script runs, regardless of
the actual configuration.

What:
Introduce AbstractDataConfigurationLoader::enableTaskType,
which removes the specified task type from the list of disabled
task types. Call that method from the refreshLinkRecommendation
maintenance script.

Note:
If link-recommendation task type is undeclared, then
the script still wouldn't work. Thanks to the work
on Community Configuration 2.0, that is impossible to happen
(even if link-recommendation is missing from the JSON file,
it would be added on the fly by merging with the defaults).

Note:
In addition to local testing, I tested the patch against
enwiki production.

Bug: T371316
Change-Id: I492ccca3d27008e0562de85e496a9124706ec369