fix(i18n): adjust broken mentorship eligibility copy
The previous copy is very misleading and causes confusing problems like
T371775.

This also moves the "mentorship turned on" setting to the top. This
makes both more sense from the overall flow as well is now consistent
with how it was in the legacy config. This now also allows us to
introduce a "Mentorship eligibility" label and more context for the
option without it being confusing to which option that belongs.
Also, not the option for Mentorship eligibility is closer to the related
input fields for minimum days and minimum edits.

This removes the message from other language files because the intend is to backport this change as soon as possible, and we don't want communities to see outdated copy.

Bug: T371775
Bug: T370318
Change-Id: I309a7f22754ff3576ec4c44d153a92010590676f