chore: maybe fix LibUp by upgrading vue and excluding coverage/ dir
Looking at the most recent failing LibUp run:
https://libraryupgrader2.wmcloud.org/logs2/1442632
there seem to be several things broken. We definitively don't need
@vue/compat anymore. Production is now squarly on Vue 3 and this is a
no-op.

Also, this adds the documentation/frontend/coverage/ directory to eslint
ignore which seems to have caused a lot of pointless linting errors in
the log.

Another change might be needed to deal with the stylelint issues.

Bug: T357224
Change-Id: I33496f3b09902675140b7a2281017285e906807b