Run fixLinkRecommendationData even when disabled in CC
The CommunityConfiguration option is intended to only disable the task
from showing up in the Homepage, not turn off all backend functionality
completely. The `AbstractDataConfigurationLoader::enableTaskType` method
was added for T371316. We probably should find a better way to do this.

Further context is that we're missing a lot of data from this script, as
can be seen on:
https://grafana.wikimedia.org/d/vGq7hbnMz/special-homepage-and-suggested-edits?orgId=1&from=1726468200000&to=1726997400000&viewPanel=38
It us unclear _why_ that data is missing. There is no obvious place for
this script to fail and no errors in any logs/output could be found so
far.

The added output commands serve two purposes:
* to provide data at least in the logs even if that data might be
  missing in Graphite
* to provide some insight into we have some viable tasks available, even
  for wikis where the community has not yet enabled this task

Getting the additional inside from this change is expected to be
particularly valuable as we change when outdated suggestions are cleared
(see Ibc178572e8b36cf4b5290915357621eabc04af77), and later will add more
suggestions when Newcomers are reading a page (see tasks related to
T376676).

Bug: T373176
Depends-On: Ie76020dc4fa3545f827e1674051530b479f01f31
Change-Id: I431d2aba14db9ab8931e21260cb2005c7276e2b8