Fix handling of undefined values after $.extend to Object.assign update
The recent update to eslint-config-wikimedia (patch 1057374)
replaced $.extend with Object.assign in several places,
including MenteeOverviewApi.js. This change introduced
a subtle but significant difference in how undefined
values are handled during object merging, causing
validation failures. An example is explained in
detail here:T372164#10057975.

In this patch it's ensured that default values are
preserved when undefined is provided in the config.
A simple unit test is created to verify that default
values are retained and defined values are applied correctly.

Bug: T372351
Change-Id: Iefff1cdf987d4969c30c65ea3fd6636bcae90f00