refactor: add proper type-hints to recently touched classes
The rev-id field is indeed sadly "mixed" because it can either be a
proper rev-id that is `?null` or it can be a string from Flow. Hopefully
we can clean up things like this once we removed that extension.

The `$timestamp` field used to be defacto `int|string` because
`wfTimestamp()` returns a string.

And since none of this was type-checked and it is via the api
user-input, all of it could have been basically any primitive type
beyond `null`. So this change introduces explicit type-checking via the
static `newFromArray` method, which is the only way that new
`QuestionRecord` objects are instantiated in production code.

Change-Id: Ic15e506ed8a5038841c72bbed2c9ea45656731cf