fix(styles): darkmode: make add a link suggestions readable
The fixed `#f4f9ff` is almost white and so the white text in darkmode is
unreadable on it.
The new color token is `#f8f9fa` and so is pretty close to the original,
and its darkmode value is `#202122`: pretty readable with white text.

Choosing this token was my call, but it is very close and so should be
fine. For now, it unblocks structured Add a Link tasks, to be used in
darkmode as well.

Bug: T373336
Bug: T369588
Change-Id: I9417ff81a72e9c048b47b86e916c892c5d6a45e7