fix: dark-mode: Make highlights in onboarding dialogues readable
`<mark>` does not seem to inherit the font-color from its parent, even
if that has higher specificity, so we need to set it explicitly.

It is not clear why we are using mix-blend-mode here in the first place,
but there seems no setting where it works well for both light- and
dark-mode, and so this unsets it in dark-mode. Though we should consider
just getting rid of mix-blend-mode entirely, this does not seem to be
its main purpose.

This change is essentially an alternative to changes
I28618281785cd9a0cb3194964f3ca937c6b9dbb4
I9fe127eef47993017d42a8a9157b5920905cb046
I0d094683702180b0d0035b37b59554714271fa06

Bug: T371564
Bug: T371146
Change-Id: Ic34f415f3efc79579690316e1146034ff5fd3521