refactor: replace less' fade() call with fixed less variable
Per documentation, all that `fade()` does is to set the transparency of
the color. A better name for that function would have been
`setColorTransparency()`. So, quite a number of the original fade(, 50)
calls were pointless because the color was already at 50% transparency.

The problem with `fade()` is that it evaluates at compile time, which
means we cannot use it with CSS custom properties, the value of which is
only available at runtime.

This now introduces a (temporary) variable to hold the original value,
which is already 50% transparent, and uses that in all the places where
previously fade was used to calculate exactly that value.

That enables us to in a follow-up commit to change
@suggestedEditsBackgroundColor to an CSS custom property that differs by
light/dark-mode.
However, since a number of our interfaces were made
compatible with dark-mode in a makeshift way by using `skin-invert
notheme` classes on them, we will have to replace
`@suggestedEditsBackgroundColorLegacyLightOnly` step by step as we switch
the respective interface to properly supporting dark-mode.

This leaves one call to less' `tint()` in T357740.

Bug: T357740
Change-Id: I55c7e00f89554131f15bc94519e6ddb9e1e13188