Suggested edits: update background color to use Codex token
This introduces a custom CSS property for the what used to be the value
for `@suggestedEditsBackgroundColor`, that is
`background-color-progressive-subtle` at 50% transparency. The dark-mode
value is the corresponding color of that token at 50% transparency as
well.

Using the .skin-invert and .notheme class-selectors in the definition of
the color allows us to use them also in places that we're currently
inverting, like the homepage.

The styles are explicitly loaded in all the places that seem like
they might use them, for lack of a better alternative.

Bug: T367986
Bug: T371146
Change-Id: Ie85b2212caccf1fa182078711da51c7472fb7d99